Improve extraction and selection of attributes.
Remove redundancies in lists of attribute names. Add select_spct_attributes() function for manipulation/edition of lists of attributes. Add accessor function to retrieve lists of names of attributes according to spectral classes. This function makes it possible to retrieve data stored in an internal list portably. Update get_attributes() methods to support easier selection of attributes to retrieve. Using a syntax similar to that used for annotations in 'ggspectra'.