Update for upcoming 'dplyr' 1.0.0
One incompatibility seems the result of a bug in dplyr::slice(). Other problems had to do with changes in the handling of attributes, which affected merge2object_spct(), and also some test cases. (Possibly a change either in 'testhat' or in R makes tests fails when the order in which attributes are stored differs between objects being compared.