Replace oper.e.generic_spctt() and update f_dispatcher()
Allowing Afr as a fully supported quantity required changes to the "oper" methods. At the same time several problems with attribute copying were detected when e2 is of class numeric. On the other hand check_spct() was not being called in some cases. (Finally replace the old and very ugly/smelly dispatch code.