Fix bugs in operators for generic.mspct
New bugs related to stripping of attributes in extract, replace and append operations on collections of spectra. Revised [, [<-, [[<- to ensure that attributes are copied. I do not know yet if they were triggered by changes in R 4.2.1, earlier changes or have been present for a long time. New test cases are still to be added.