Fix nasty bug in rbindspct()
The bug in handling of the idfactor in the where.measured attribute data frame was breaking among other things trim_wl() and clip_wl() applied to spectra in long form, and thus also breaking the use of range in autoplot in ggspectra, and most likely in some other methods. Added test cases for source_spct as the code is shared by all the derived classes.