Rewrite find_peaks(), add find_valleys()
valleys() methods were broken in 0.13.0. Updated to use find_valleys() as the thresholds and related parameters need to be modified to work as expected with find_peaks(-x). Add also better error messages and checks of validity for arguments passed to find_peaks(). Add unit tests and check once again. The new logic is now working in existing and new cases for peaks and valleys for source_spct objects.